# sorter/leveldb(ticdc): optimize write and read throughput