--- a/pkg/cmd/server/server_test.go
+++ b/pkg/cmd/server/server_test.go
@@ -180,8 +180,8 @@ func TestParseCfg(t *testing.T) {
 			EnableTableActor: false,
 			EnableDBSorter:   false,
 			DB: &config.DBConfig{
-				Count:                       16,
-				Concurrency:                 256,
+				Count:                       8,
+				Concurrency:                 128,
 				MaxOpenFiles:                10000,
 				BlockSize:                   65536,
 				BlockCacheSize:              4294967296,
@@ -456,8 +456,8 @@ cert-allowed-cn = ["dd","ee"]
 			EnableTableActor: false,
 			EnableDBSorter:   false,
 			DB: &config.DBConfig{
-				Count:                       16,
-				Concurrency:                 256,
+				Count:                       8,
+				Concurrency:                 128,
 				MaxOpenFiles:                10000,
 				BlockSize:                   65536,
 				BlockCacheSize:              4294967296,